+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - improved method naming in conversation handler api
 - replace with ConversationHandler `WithDefaultDefaultResponse` method with `WithDefaultDefaultSequence`
 - added context parameter to sequence process method
+- added context and update parameters to sequence start method
 
 ## v0.9.2